backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "radius-ui-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn[standard]==0.34.0",
    "sqlalchemy==2.0.48",
    "asyncpg==0.30.0",
    "alembic==1.14.1",
    "pydantic==2.11.3",
    "pydantic-settings==2.9.0",
    "PyJWT==2.10.1",
    "pwdlib[argon2]==0.2.1",
    "python-multipart==0.0.20",
    "httpx==0.28.1",
    "psycopg2-binary==2.9.10",
    "email-validator==2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "pytest-asyncio==0.25.3",
    "pytest-cov==6.1.0",
    "ruff==0.11.4",
    "aiosqlite==0.21.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "alembic*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
